split LED from main file
