+++ .env
@@ -1,0 +1,2 @@
+Discord_Token =
+Discord_ServerID =
