--- .github/ISSUE_TEMPLATE/custom.md
@@ -1,10 +1,0 @@
----
-name: Custom issue template
-about: Describe this issue template's purpose here.
-title: ''
-labels: ''
-assignees: ''
-
----
-
-
