+++ cogs/LED_OnOff.py
@@ -3,29 +3,9 @@ import gpiozero
 import discord
 import os
 import re
-from dotenv import load_dotenv
-
-#loading variables from .env which contains auth tokens
-load_dotenv()
-TOKEN = os.getenv('Discord_Token')
-GUILD = os.getenv('Discord_Server')
-GUILDID = os.getenv('Discord_ServerID')
-
-#starts the discord client
-client = discord.Client()
 
 #maps GPIO pin to variable. active_high property make it use the 3.3v and GPIO pin
 redLED = gpiozero.LED(17, active_high=False)
-
-#event handler on start
-@client.event
-async def on_ready():
-    guild = discord.utils.get(client.guilds, id=int(GUILDID))
-
-    print(
-        f'{client.user} has connected to the following server:\n'
-        f'{guild.name}(id: {guild.id})'
-        )
 
 #event handler for when a message is recieved
 @client.event
@@ -48,6 +28,3 @@ async def on_message(message):
                 return
         else:
             return
-
-#starts the client and runs the script with the provided token in .env file
-client.run(TOKEN)
